Adding amr claim in OIDC id_token
The amr claim is an optional id_token claim specified in rfc8176 (https://datatracker.ietf.org/doc/html/rfc8176). It stands for authentication method reference and specifies the exact methods of authentication used to verify a user. A client can subsequently access this information from the returned id_token and make use of it if they wish. For example, if a client wishes to ensure that a user authenticates with more than one factor of authentication, they can check the length of the amr claim. If it is greater than or equal to 2, then they know that multiple factors were used. The exact factors are also specified, for example pwd for username and password and otp for one-time password.

Currently, only two values can be present in the amr claim: pwd and otp. At present, this is enough to cover all methods of authentication that are currently available, as the only additional method of authentication available so far is an authenticator app, denoted by "otp".

MFA users currently can't authenticate to use a client, as they always get redirected to the IAM dashboard (this was for demo purposes but is now outdated). This will be fixed in a commit shortly following this one.

This commit also includes an additional migration to the database. It adds a table to store the last authentication method references used by an account when they last signed in. This isn't ideal, as this information should only be temporary and as such stored in an authentication object. However, I could not find a way to implement this (by extending either of the authentication or user details objects) but may figure this out later on. On the plus side, this is a working solution.