Adding SQL migration files for MFA
This includes a new table for multi-factor secrets, a new table for recovery codes associated with a multi-factor secret and changes to the iam_account table that link a multi-factor secret to an account.

In a separate migration, we are also adding a new authority called ROLE_PRE_AUTHENTICATED. This is assigned to a user if they have MFA enabled on their account and they complete one factor of authentication . The role is used to grant the user access only to pages where they can authenticate through an additional factor of authentication.