OAuth2 client login flow now works for MFA users
This involved refactoring the way TOTPs were verified. Previously, this was done through the controller for the verification page itself. This logic is being moved into a dedicated authentication provider, executed by a dedicated authentication processing filter on a new authentication entry point.

Still to test failed TOTP entries and account for recovery codes as well